backend/foodgram/requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
class-registry==2.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2.4
django-autoslug==1.9.8
django-colorfield==0.7.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
drf-yasg==1.21.4
filters==1.3.2
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.7.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.8.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.3
pdfkit==0.6.1
pep8-naming==0.11.1
Pillow==9.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.17.1
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
regex==2022.6.2
reportlab==3.6.11
requests==2.26.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.6
zipp==3.8.0
